Add Channel Actions to ActionCable guide [ci skip]
Actions defined on Channels were not explained in the ActionCable guide,
including what methods are callable and how routing works.

This commit copies the documentation from
actioncable/lib/actioncable/channels/base.rb with small modifications to
make it fit in the guides.

The unsubscribed method is added in the example above to make it's
appearance in the actions example less confusing.