Add `ActionDispatch::HostAuthorization` to "Configuring Middleware" guide [ci skip]
This commit copies info from
https://github.com/rails/rails/pull/33145
([an excellent changelog](https://github.com/rails/rails/blob/6-0-stable/railties/CHANGELOG.md)) to
our guides.

Closes https://github.com/rails/rails/issues/36959
Not sure whether there is a need mentioning `config.hosts` in the "Upgrading to Rails 6.0" guide
since it is configured to work in the development environment by default and we guide how to deal
with "blocked host" issue, see:
https://github.com/rails/rails/blob/6-0-stable/actionpack/lib/action_dispatch/middleware/templates/rescues/blocked_host.html.erb

I would like to backport this to `6-0-stable` so users will be able to
find info about `ActionDispatch::HostAuthorization` middleware and how
to configure it.

Give credit to @gsamokovarov since I copied the info from the changelog. <3

[bogdanvlviv, Genadi Samokovarov]