Prevent `has_one` `build_association` from `touch` parent record if the record isn't committed
Previously, if `build_association` was called multiple times for a
`has_one` association but never committed to the database, the first
newly-associated record would trigger `touch` during the attempted
removal of the record.

For example:

    class Post < ActiveRecord::Base
      has_one :comment, inverse_of: :post, dependent: :destroy
    end

    class Comment < ActiveRecord::Base
      belongs_to :post, inverse_of: :comment, touch: true
    end

    post = Post.create!
    comment_1 = post.build_comment
    comment_2 = post.build_comment

When `comment_2` is initialized, the `has_one` would attempt to destroy
`comment_1`, triggering a `touch` on `post` from an association record
that hasn't been committed to the database.

This checks if the association is persisted, and if it isn't, uses
`NoTouching` to prevent the owner from being touched during
`build_association`.