Deprecate marshal behavior on schema cache
YAML has been used to serialize the schema cache ever since 2016 with
Rails 5.1: 4c00c6e.

The marshaling behavior was removed in March 2019 (65f2eea), and then reintroduced in
e5f1b363 the next day.

There was no discussion about this, so it's not clear what broke or why the marshal
behavior is still necessary.