Improve zeitwerk:check warning message.
The messages that are printed to the console when users run rails zeitwerk:check
can be a little confusing if they've generated a mailer, as the test files aren't
included by default, and a warning appears about this.

To address this issue the warning message has been improved so that it is clearer
that this behaviour is normal and it explicitly states that the user must configure
eager loading to silence these warnings. Fixes #38156