Prevent infinite recursion in Object.as_json
The fallback implementation for `as_json` on Object just dumps the instance variables as name value pairs in a Hash. However, this is very susceptible to infinite recursion when dumping an object that maintains references to other objects that then maintain back references to the original object.

So, for example, a tree structure where the parent has a list of the children and each child has a reference to it's parent results in a `SystemStackError` when calling as_json on any node.

The fix is to maintain a state of the current stack of objects being used to construct the as_json hash. If an object is already being used, it is left out of the hash in lieu of having it raise a stack level too deep error.