Load schema cache for all database connections
Rather than loading only the schema cache for the primary database, loop through
all the possible configurations and load the schema cache for each of them.

Note that this only loads the cache onto the default connection pools. If
any other connections are established outside of this, then they will not
get their schema cache loaded.