Clarify passing an array as a cookie's domain
When working through an issue, I found myself a little confused stumbling upon some code like this:

```rb
cookies[...] = {
  # ...
  domain: %w(.site.io .site.com)
}
```

In my mind I was thinking "is this some weird part of the cookie spec? or is this Rails?" and the docs didn't offer much insight into how it works. Going back into the commit where that line was added, the folks who added it were confused to until someone linked to the source.

I figured that clarifying this in the docs would be useful. Of course, people should know how cookies work before fiddling with them (unlike me), but since cookies are so sensitive, I think the docs should opt for being explicit about this.

Co-Authored-By: Norm Provost <norman.provost@intercom.io>