Fix load_additional_types handling of large OIDs
In Postgres, OIDs are *unsigned* 4 byte integers. Postgres allows
casting oids to integers so that you can have access to more operators;
however, doing this could lead to some issues if the oids are very large
as they might around to the negative number range.

From the Postgres docs:

> The oid type itself has few operations beyond comparison. It can be
> cast to integer, however, and then manipulated using the standard
> integer operators. (Beware of possible signed-versus-unsigned
> confusion if you do this.)

In most real-world databases this is never going to be a problem. In
order to reach the overflow, you have to have 2,147,483,648 objects in
your database (not rows, but relations, types, views, etc.). In extreme
cases though it is possible to reach that number and encounter some
buggy behavior.

See: https://www.postgresql.org/docs/12/datatype-oid.html