Callbacks conditionals and Symbol#to_proc syntax
Callback creation checks arity to determine which kind of conditional to build. In the case of a normal block that accepts one argument, as in:

```
Proc.new { |record| record.errors }
```

you end up with arity 1. But in case you convert that into using Symbol#to_proc syntax, as in:

```
Proc.new(&:errors)
```

you end up with arity -1. This will fail with an error of no receiver given, because it thinks a block was passed to the callback. Effectively, these two should amount to the same thing, but because of the nature of the metaprogramming happening here, we end up with an error.

There are a couple of ways to determine whether or not a Proc is using this kind of syntax, but the easiest way it to just check Proc#to_s and match against a regex. Other options include:

* using the `source_location` method to determine the original place of the block creation and parsing the source using something from `RubyVM`
* calling the Proc with different numbers of arguments and rescuing to determine the correct number

The Proc#to_s feels like the simplest solution of these 3.