Generate `aria-label` in calls to `rich_text_area`
Prior to this commit, calls to `rich_text_area` would generate
`<trix-editor>` elements that are largely inaccessible by default.

When either parent `<label>` elements or `<label
for="rich-text-field-id">` receive focus, the underlying `<trix-editor>`
elements do not receive focus in the same way that browser-default
fields like `<input>` and `<textarea>` do.

When a `<trix-editor>` that lacks an `aria-label` attribute receives
focus, accessibility tools cannot detect its appropriate label text.

The newly introduced [`fill_in_rich_text_area`][fill_in_rich_text_area]
will [resolve its `locator` argument based on an element's
`aria-label`][locator-aria-label]. This is going to be a helpful tool to
ensure that tests can add coverage in a way that depends on the presence
of the `aria-label` field.

This commit takes that idea a step further, and extends the
`rich_text_area` element to accept the `label: true` option, which
generates an appropriate `aria-label` value in the same way that a call
to `form.label(:attribute)` would.

Ideally, this generation would happen by default, but for the sake of
backwards compatibility, the behavior is opt-in instead of opt-out.

If an `aria-label` attribute is passed to the `rich_text_field`, the
`label` option is ignored.

In either case, the `label` option is never passed along to the
HTML that the `rich_text_area` generates.

[fill_in_rich_text_area]: https://github.com/rails/rails/blob/0105fd4a1ec5579fbc824c7a864794ba90351c39/actiontext/lib/action_text/system_test_helper.rb#L5-L29
[locator-aria-label]: https://github.com/rails/rails/blob/0105fd4a1ec5579fbc824c7a864794ba90351c39/actiontext/lib/action_text/system_test_helper.rb#L44